Fix types breaking compilation

+++ src/model/modeling_video_mask2former.py
@@ -820,7 +820,7 @@ def multi_scale_deformable_attention(
 ) -> Tensor:
     batch_size, _, num_heads, hidden_dim = value.shape
     _, num_queries, num_heads, num_levels, num_points, _ = sampling_locations.shape
-    value_list = value.split([height * width for height, width in value_spatial_shapes], dim=1)
+    value_list = value.split([int(height * width) for height, width in value_spatial_shapes], dim=1)
     sampling_grids = 2 * sampling_locations - 1
     sampling_value_list = []
     for level_id, (height, width) in enumerate(value_spatial_shapes):
@@ -1403,9 +1403,9 @@ class VideoMask2FormerPixelDecoder(nn.Module):
         split_sizes = [None] * self.num_feature_levels
         for i in range(self.num_feature_levels):
             if i < self.num_feature_levels - 1:
-                split_sizes[i] = level_start_index[i + 1] - level_start_index[i]
+                split_sizes[i] = int(level_start_index[i + 1] - level_start_index[i])
             else:
-                split_sizes[i] = last_hidden_state.shape[1] - level_start_index[i]
+                split_sizes[i] = int(last_hidden_state.shape[1] - level_start_index[i])
 
         encoder_output = torch.split(last_hidden_state, split_sizes, dim=1)
